requirements.txt
python-glanceclient==2.16.0
keystoneauth1==3.13.1
